refactor(ast): remove overlapping and mislabeled abstractions (#3573)
Epic 2:
- 2a: header comments on dedupe.ts and utils/fileMerge.ts drawing the
  schema-shape vs file-member dedup boundary. No logic change.
- 2b: move syncSchemaRef (a SchemaNode -> SchemaNode transform) into
  transformers.ts. It stays exported from @kubb/ast/utils.
- 2c: stop surfacing createOperationParams through the factory namespace and
  expose it via @kubb/ast/utils. Move the OpenAPI discriminator helpers
  createDiscriminantNode and findDiscriminator out of @kubb/ast into
  @kubb/adapter-oas (their only consumer), keeping @kubb/ast spec-agnostic.
