+++ packages/ast/src/utils/index.ts
@@ -1,8 +1,9 @@
 export { isValidVarName } from '@internals/utils'
 export { buildJSDoc, buildList, buildObject, objectKey } from './codegen.ts'
 export { extractStringsFromNodes } from './extractStringsFromNodes.ts'
-export { childName, enumPropName, extractRefName, findDiscriminator, isStringType, resolveGroupType, syncSchemaRef } from './refs.ts'
+export { childName, enumPropName, extractRefName, isStringType, resolveGroupType } from './refs.ts'
+export { syncSchemaRef } from '../transformers.ts'
 export { getNestedAccessor, jsStringEscape, stringify, stringifyObject, toRegExpString, trimQuotes } from './strings.ts'
 export { collectUsedSchemaNames, containsCircularRef, findCircularSchemas } from './schemaGraph.ts'
-export { buildGroupParam, buildTypeLiteral, caseParams, resolveParamType } from './operationParams.ts'
+export { buildGroupParam, buildTypeLiteral, caseParams, createOperationParams, resolveParamType } from './operationParams.ts'
 export type { BuildGroupArgs, ParamGroupType } from './operationParams.ts'
