+++ packages/ast/src/utils/refs.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import { createSchema } from '../nodes/schema.ts'
-import { childName, enumPropName, extractRefName, findDiscriminator, isStringType, resolveRefName, syncSchemaRef } from './refs.ts'
+import { childName, enumPropName, extractRefName, isStringType, resolveRefName } from './refs.ts'
 
 describe('extractRefName', () => {
   it('extracts schema name from component schema refs', () => {
@@ -57,37 +57,6 @@ describe('enumPropName', () => {
   })
 })
 
-describe('findDiscriminator', () => {
-  it('returns the mapping key for a matching discriminator ref', () => {
-    const mapping = {
-      cat: '#/components/schemas/Cat',
-      dog: '#/components/schemas/Dog',
-    }
-
-    expect(findDiscriminator(mapping, '#/components/schemas/Dog')).toBe('dog')
-  })
-
-  it.each([
-    {
-      label: 'mapping is missing',
-      mapping: undefined,
-      ref: '#/components/schemas/Dog' as string | undefined,
-    },
-    {
-      label: 'ref is missing',
-      mapping: { cat: '#/components/schemas/Cat' },
-      ref: undefined,
-    },
-    {
-      label: 'ref does not match any mapping entry',
-      mapping: { cat: '#/components/schemas/Cat' },
-      ref: '#/components/schemas/Dog',
-    },
-  ])('returns null when $label', ({ mapping, ref }) => {
-    expect(findDiscriminator(mapping, ref)).toBeNull()
-  })
-})
-
 describe('isStringType', () => {
   it('returns true for plain string-like schema types', () => {
     expect(isStringType(createSchema({ type: 'string' }))).toBe(true)
@@ -106,38 +75,5 @@ describe('isStringType', () => {
     expect(isStringType(createSchema({ type: 'date', representation: 'date' }))).toBe(false)
     expect(isStringType(createSchema({ type: 'time', representation: 'date' }))).toBe(false)
     expect(isStringType(createSchema({ type: 'number' }))).toBe(false)
-  })
-})
-
-describe('syncSchemaRef', () => {
-  it('returns a merged schema for a ref node that has a resolved schema', () => {
-    const resolved = createSchema({ type: 'object' })
-    const ref = createSchema({
-      type: 'ref',
-      name: 'Pet',
-      ref: '#/components/schemas/Pet',
-      schema: resolved,
-    })
-
-    const merged = syncSchemaRef(ref)
-    expect(merged).not.toBeNull()
-    expect(merged?.type).toBe('object')
-  })
-
-  it('returns a merged schema with sibling overrides applied over the resolved schema', () => {
-    const resolved = createSchema({ type: 'object', description: 'Original' })
-    const ref = createSchema({
-      type: 'ref',
-      name: 'Pet',
-      ref: '#/components/schemas/Pet',
-      schema: resolved,
-      description: 'Override',
-      readOnly: true,
-    })
-
-    const merged = syncSchemaRef(ref)
-    expect(merged?.description).toBe('Override')
-    expect(merged?.readOnly).toBe(true)
-    expect(merged?.type).toBe('object')
   })
 })
